Move add_display_name to utils.py
